Update of Bluewing Client to build 88, Server to build 20
Client + Server:
• Updated DarkEdif SDK to v2 (better update notifs/fixed editor display).

Server (this commit):
• Fixed some expression error checks using wrong variable size

Since I didn't originally notice...
Server (commit 08a901341a102af790f1b57b5b9ea6d0150892eb):
• Fixed bug with reading messages when messagereader::getremaining() is called.
• Fixed kicking clients when a client->server message with no data (beside subchannel/variant) was sent; server was mistakenly set to not accept empty messages, and due to it seeing a message corruption, it kicked the client immediately.
• Fixed crash caused by lock being freed twice during some message parsing errors.
• Also switched kicks for broken messages from clean exit to immediate exit. It may not be safe to wait for connection to finish processing, as client could be pouring messages into their end.

+++ DarkEdif/Bluewing Server/Expressions.cpp
@@ -127,7 +127,7 @@ unsigned int Extension::UnsignedInteger(int index)
 {
 	if (index < 0)
 		return CreateError("Could not read unsigned integer from received binary, index %i is less than 0.", index), 0;
-	if (threadData->receivedMsg.content.size() - index < sizeof(signed short))
+	if (threadData->receivedMsg.content.size() - index < sizeof(unsigned int))
 		return CreateError("Could not read unsigned integer from received binary at position %i, amount of message remaining is smaller than variable to be read.", index), 0;
 
 	return (*(unsigned int *)(threadData->receivedMsg.content.data() + index));
@@ -136,7 +136,7 @@ int Extension::SignedInteger(int index)
 {
 	if (index < 0)
 		return CreateError("Could not read signed integer from received binary, index %i is less than 0.", index), 0;
-	if (threadData->receivedMsg.content.size() - index < sizeof(signed short))
+	if (threadData->receivedMsg.content.size() - index < sizeof(int))
 		return CreateError("Could not read signed integer from received binary at position %i, amount of message remaining is smaller than variable to be read.", index), 0;
 
 	return (*(int *)(threadData->receivedMsg.content.data() + index));
@@ -145,7 +145,7 @@ float Extension::Float(int index)
 {
 	if (index < 0)
 		return CreateError("Could not read float from received binary, index %i is less than 0.", index), 0.0f;
-	if (threadData->receivedMsg.content.size() - index < sizeof(signed short))
+	if (threadData->receivedMsg.content.size() - index < sizeof(float))
 		return CreateError("Could not read float from received binary at position %i, amount of message remaining is smaller than variable to be read.", index), 0.0f;
 
 	return (*(float *)(threadData->receivedMsg.content.data() + index));
